Add option to immediately execute
This mode is if the return value of the captured object is needed for
further logic (i.e. not just a pure side effect call like
CanvasRenderingContext2D). Use this mode to stop deferring, using
deferred to just capture calls for future examination.

Use by passing in immediatelyExecute: true in the options object as the
2nd argument.